+++ src/pytakes/run/runner.py
@@ -1,6 +1,8 @@
 from datetime import datetime
 from typing import List
 
+from loguru import logger
+
 from pytakes import MinerCollection, SentenceBoundary
 from pytakes.corpus import get_next_from_corpus
 from pytakes.dict.textitem import process_textitem
@@ -8,11 +10,13 @@ from pytakes.run.schema import validate_config
 from pytakes.run.simple_run import load_keywords, load_negation, output_context_manager
 
 
-def run(corpus=None, output=None, keywords: List = None, negation=None, logger=None):
+def run(corpus=None, output=None, keywords: List = None, negation=None, log=None):
     """
 
     :return:
     """
+    if log and log.get('file', None):
+        logger.add(log['file'], level=log.get('level', 'DEBUG'))
     mc = MinerCollection(ssplit=SentenceBoundary().ssplit)
     mc.add(load_keywords(*keywords))
     mc.add(load_negation(**negation))
@@ -21,10 +25,12 @@ def run(corpus=None, output=None, keywords: List = None, negation=None, logger=N
         'metalabels': ['file'],
     }
     with output_context_manager(**{**default_output, **output}) as out:
-        for i, ti in enumerate(get_next_from_corpus(**corpus)):
+        for i, ti in enumerate(get_next_from_corpus(**corpus), start=1):
             for results, sent in process_textitem(ti, mc):
                 for result in results:
                     out.writerow(result, meta=list(ti.meta.values()), text=sent)
+            if i % 10000 == 0:
+                logger.info(f'Completed processing of {i} records.')
 
 
 def run_config(config_file):